requirements.txt
certifi==2019.11.28
configparser==4.0.2
joblib==0.14.1
minio==5.0.7
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
urllib3==1.25.8
keras==2.3.1
tensorflow==2.1.0
psutil
pillow